1. fix bug of mix_DMk_2D.set_mixing_beta(CHR_MIX.get_mixing_beta()), when CHR_MIX.autoset==true 2. change default value
