+++ source/module_ri/Mix_DMk_2D.h
@@ -15,65 +15,69 @@
 class Mix_DMk_2D
 {
 public:
-/**
-     * @brief Sets the number of k-points and gamma_only flag.
-     * @param nks Number of k-points.
-     * @param gamma_only_in Flag indicating if only gamma point is mixed.
-     * @return Reference to the current object.
-     */
+	/**
+	 * @brief Sets the number of k-points and gamma_only flag.
+	 * @param nks Number of k-points.
+	 * @param gamma_only_in Flag indicating if only gamma point is mixed.
+	 * @return Reference to the current object.
+	 */
 	Mix_DMk_2D &set_nks(const int nks, const bool gamma_only_in);
-/**
-     * @brief Sets the mixing mode.
-     * @param mixing_mode Mixing mode.
-     * @return Reference to the current object.
-     */
-	Mix_DMk_2D &set_mixing_mode(const Mixing_Mode mixing_mode);
-    Mix_DMk_2D& set_mixing_beta(const double mixing_beta);
-    template<typename ChgMix>
-    Mix_DMk_2D& set_coef_pulay(const int iter, const ChgMix& chr_mix);
 
-/**
-     * @brief Mixes the density matrix.
-     * @param dm double Density matrix.
-     * @param flag_restart Flag indicating if it is a restart.
-     */
+	/**
+	 * @brief Sets the mixing mode.
+	 * @param mixing_mode Mixing mode.
+	 * @return Reference to the current object.
+	 */
+	Mix_DMk_2D &set_mixing_mode(const Mixing_Mode mixing_mode);
+
+	/**
+	 * @brief Sets the mixing beta.
+	 * @param mixing_beta Mixing beta.
+	 * @return Reference to the current object.
+	 */
+	Mix_DMk_2D &set_mixing_beta(const double mixing_beta);
+
+	/**
+	 * @brief Sets the pulay mixing coefficients from class Charge_Mixing.
+	 * @tparam ChgMix Type of the charge mixing coefficient.
+	 * @param iter Iteration step, start from 1.
+	 * @param chr_mix Object of Charge_Mixing.
+	 * @return Reference to the current object.
+	 */
+	template<typename ChgMix>
+	Mix_DMk_2D &set_coef_pulay(const int iter, const ChgMix& chr_mix);
+
+	/**
+	 * @brief Mixes the double density matrix.
+	 * @param dm Double Density matrix.
+	 * @param flag_restart Flag indicating whether restart mixing.
+	 */
 	void mix(const std::vector<ModuleBase::matrix> &dm, const bool flag_restart);
 
-/**
-     * @brief Mixes the complex density matrix.
-     * @param dm Complex density matrix.
-     * @param flag_restart Flag indicating if it is a restart.
-     */
+	/**
+	 * @brief Mixes the complex density matrix.
+	 * @param dm Complex density matrix.
+	 * @param flag_restart Flag indicating whether restart mixing.
+	 */
 	void mix(const std::vector<ModuleBase::ComplexMatrix> &dm, const bool flag_restart);
 
-/**
-     * @brief Returns the gamma density matrix.
-     * @return Vector of pointers to gamma density matrices.
-     */
+	/**
+	 * @brief Returns the double density matrix.
+	 * @return Double density matrices for each k-points.
+	 */
 	std::vector<const ModuleBase::matrix*> get_DMk_gamma_out() const;
-/**
-     * @brief Returns the k-point density matrix.
-     * @return Vector of pointers to k-point density matrices.
-     */
+	/**
+	 * @brief Returns the complex density matrix.
+	 * @return Complex density matrices for each k-points.
+	 */
 	std::vector<const ModuleBase::ComplexMatrix*> get_DMk_k_out() const;
 
 private:
 	std::vector<Mix_Data<ModuleBase::matrix>> mix_DMk_gamma;
-    std::vector<Mix_Data<ModuleBase::ComplexMatrix>> mix_DMk_k;
-    bool gamma_only;
+	std::vector<Mix_Data<ModuleBase::ComplexMatrix>> mix_DMk_k;
+	bool gamma_only;
 };
 
-template<typename ChgMix>
-Mix_DMk_2D& Mix_DMk_2D::set_coef_pulay(const int iter, const ChgMix& chr_mix)
-{
-	ModuleBase::TITLE("Mix_DMk_2D","set_coef_pulay");
-	if(gamma_only)
-		for(Mix_Data<ModuleBase::matrix> &mix_one : this->mix_DMk_gamma)
-			mix_one.set_coef_pulay(iter, chr_mix);
-	else
-		for(Mix_Data<ModuleBase::ComplexMatrix> &mix_one : this->mix_DMk_k)
-			mix_one.set_coef_pulay(iter, chr_mix);
-	return *this;
-}
+#include "Mix_DMk_2D.hpp"
 
 #endif
